requirements.txt
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
datasets
sentencepiece
sacrebleu
transformers
deepspeed==0.13.1
accelerate==0.27.2
tokenizers==0.19.0
unbabel-comet
huggingface-hub 
langcodes
